+++ src/Query.php
@@ -11,29 +11,29 @@ class Query
 
     protected PDO $pdo; // Instance
 
-    /*
-     * Possible $query keys include:
-     *  - from
-     *  - distinct
-     *  - inner_join
-     *  - left_join
-     *  - right_join
-     *  - columns
-     *  - where
-     *  - sort
-     *  - limit
-     *  - offset
-     */
-
-    private array $query = [];
-
-    private array $placeholders = [];
-
     public function __construct(PDO $pdo)
     {
         $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION); // Throw exceptions
         $this->pdo = $pdo;
     }
+
+    /*
+     * $query array keys
+     */
+    private const QUERY_FROM = 'from';
+    private const QUERY_DISTINCT = 'distinct';
+    private const QUERY_INNER_JOIN = 'inner_join';
+    private const QUERY_LEFT_JOIN = 'left_join';
+    private const QUERY_RIGHT_JOIN = 'right_join';
+    private const QUERY_COLUMNS = 'columns';
+    private const QUERY_WHERE = 'where';
+    private const QUERY_SORT = 'sort';
+    private const QUERY_LIMIT = 'limit';
+    private const QUERY_OFFSET = 'offset';
+
+    private array $query = [];
+
+    private array $placeholders = [];
 
     private string $table = '';
 
@@ -44,7 +44,7 @@ class Query
      */
     public function table(string $table): self
     {
-        $this->query['from'] = ' FROM ' . $table;
+        $this->query[self::QUERY_FROM] = ' FROM ' . $table;
 
         $this->table = $table;
 
@@ -58,7 +58,7 @@ class Query
      */
     public function distinct(): self
     {
-        $this->query['distinct'] = 'DISTINCT ';
+        $this->query[self::QUERY_DISTINCT] = 'DISTINCT ';
         return $this;
     }
 
@@ -72,7 +72,7 @@ class Query
      */
     public function innerJoin(string $table, string $col1, string $col2): self
     {
-        $this->query['inner_join'] = ' INNER JOIN ' . $table . ' ON ' . $col1 . ' = ' . $col2;
+        $this->query[self::QUERY_INNER_JOIN] = ' INNER JOIN ' . $table . ' ON ' . $col1 . ' = ' . $col2;
         return $this;
     }
 
@@ -86,7 +86,7 @@ class Query
      */
     public function leftJoin(string $table, string $col1, string $col2): self
     {
-        $this->query['left_join'] = ' LEFT JOIN ' . $table . ' ON ' . $col1 . ' = ' . $col2;
+        $this->query[self::QUERY_LEFT_JOIN] = ' LEFT JOIN ' . $table . ' ON ' . $col1 . ' = ' . $col2;
         return $this;
     }
 
@@ -100,7 +100,7 @@ class Query
      */
     public function rightJoin(string $table, string $col1, string $col2): self
     {
-        $this->query['right_join'] = ' RIGHT JOIN ' . $table . ' ON ' . $col1 . ' = ' . $col2;
+        $this->query[self::QUERY_RIGHT_JOIN] = ' RIGHT JOIN ' . $table . ' ON ' . $col1 . ' = ' . $col2;
         return $this;
     }
 
@@ -137,7 +137,7 @@ class Query
 
             }
 
-            $this->query['columns'][] = $column;
+            $this->query[self::QUERY_COLUMNS][] = $column;
 
         }
 
@@ -245,10 +245,32 @@ class Query
 
     }
 
+    /*
+     * Operators
+     */
+    public const OPERATOR_EQUALS = 'eq';
+    public const OPERATOR_DOES_NOT_EQUAL = '!eq';
+    public const OPERATOR_LESS_THAN = 'lt';
+    public const OPERATOR_GREATER_THAN = 'gt';
+    public const OPERATOR_LESS_THAN_OR_EQUAL = 'le';
+    public const OPERATOR_GREATER_THAN_OR_EQUAL = 'ge';
+    public const OPERATOR_STARTS_WITH = 'sw';
+    public const OPERATOR_DOES_NOT_START_WITH = '!sw';
+    public const OPERATOR_ENDS_WITH = 'ew';
+    public const OPERATOR_DOES_NOT_END_WITH = '!ew';
+    public const OPERATOR_HAS = 'has';
+    public const OPERATOR_DOES_NOT_HAVE = '!has';
+    public const OPERATOR_IN = 'in';
+    public const OPERATOR_NOT_IN = '!in';
+    public const OPERATOR_NULL = 'null';
+
+    public const VALUE_TRUE = 'true';
+    public const VALUE_FALSE = 'false';
+
     /**
      * Adds a WHERE clause to the query.
      *
-     * If the column type is json, keys from within the JSON string can be searched with the format of COLUMN->KEY.
+     * If the column type is JSON, keys from within the JSON string can be searched with the format of COLUMN->KEY.
      * JSON fields which do not exist are treated as null.
      *
      * Available operators are:
@@ -269,7 +291,15 @@ class Query
      * - !in (not in)
      * - null (is or is not null)
      *
+     * The OPERATOR_* constants can be used for this purpose.
+     *
+     * The in and !in operators accept multiple comma-separated values.
+     *
      * The "null" operator accepts two values: true and false for is null or is not null.
+     * The VALUE_* constants can be used for this purpose.
+     *
+     * NOTE: Some native MySQL functions can be used as the $value, however, they will be
+     * injected into the query as strings, so they can be vulnerable to SQL injection.
      *
      * @param string $column
      * @param string $operator
@@ -280,39 +310,39 @@ class Query
     public function where(string $column, string $operator, mixed $value): self
     {
 
-        if (!isset($this->query['where'])) {
-            $this->query['where'] = ' WHERE ';
+        if (!isset($this->query[self::QUERY_WHERE])) {
+            $this->query[self::QUERY_WHERE] = ' WHERE ';
         } else {
-            $this->query['where'] .= ' AND ';
+            $this->query[self::QUERY_WHERE] .= ' AND ';
         }
 
         if (!in_array($operator, [
-            'eq',
-            '!eq',
-            'lt',
-            'gt',
-            'le',
-            'ge',
-            'sw',
-            '!sw',
-            'ew',
-            '!ew',
-            'has',
-            '!has',
-            'in',
-            '!in',
-            'null'
+            self::OPERATOR_EQUALS,
+            self::OPERATOR_DOES_NOT_EQUAL,
+            self::OPERATOR_LESS_THAN,
+            self::OPERATOR_GREATER_THAN,
+            self::OPERATOR_LESS_THAN_OR_EQUAL,
+            self::OPERATOR_GREATER_THAN_OR_EQUAL,
+            self::OPERATOR_STARTS_WITH,
+            self::OPERATOR_DOES_NOT_START_WITH,
+            self::OPERATOR_ENDS_WITH,
+            self::OPERATOR_DOES_NOT_END_WITH,
+            self::OPERATOR_HAS,
+            self::OPERATOR_DOES_NOT_HAVE,
+            self::OPERATOR_IN,
+            self::OPERATOR_NOT_IN,
+            self::OPERATOR_NULL
         ])) {
             throw new QueryException('Unable to build query: invalid operator (' . $operator . ') for column (' . $column . ')');
         }
 
         $operator = str_replace([
-            'eq',
-            '!eq',
-            'lt',
-            'gt',
-            'le',
-            'ge'
+            self::OPERATOR_EQUALS,
+            self::OPERATOR_DOES_NOT_EQUAL,
+            self::OPERATOR_LESS_THAN,
+            self::OPERATOR_GREATER_THAN,
+            self::OPERATOR_LESS_THAN_OR_EQUAL,
+            self::OPERATOR_GREATER_THAN_OR_EQUAL
         ], [
             '=',
             '!=',
@@ -328,76 +358,76 @@ class Query
 
         switch ($operator) {
 
-            case 'sw':
+            case self::OPERATOR_STARTS_WITH:
 
                 if ($this->is_function($value)) {
-                    $this->query['where'] .= $column . ' LIKE ' . $value;
+                    $this->query[self::QUERY_WHERE] .= $column . ' LIKE ' . $value;
                     break;
                 }
 
                 $this->placeholders[] = $value . '%';
-                $this->query['where'] .= $column . ' LIKE ?';
+                $this->query[self::QUERY_WHERE] .= $column . ' LIKE ?';
                 break;
 
-            case '!sw':
+            case self::OPERATOR_DOES_NOT_START_WITH:
 
                 if ($this->is_function($value)) {
-                    $this->query['where'] .= $column . ' NOT LIKE ' . $value;
+                    $this->query[self::QUERY_WHERE] .= $column . ' NOT LIKE ' . $value;
                     break;
                 }
 
                 $this->placeholders[] = $value . '%';
-                $this->query['where'] .= $column . ' NOT LIKE ?';
+                $this->query[self::QUERY_WHERE] .= $column . ' NOT LIKE ?';
                 break;
 
-            case 'ew':
+            case self::OPERATOR_ENDS_WITH:
 
                 if ($this->is_function($value)) {
-                    $this->query['where'] .= $column . ' LIKE ' . $value;
+                    $this->query[self::QUERY_WHERE] .= $column . ' LIKE ' . $value;
                     break;
                 }
 
                 $this->placeholders[] = '%' . $value;
-                $this->query['where'] .= $column . ' LIKE ?';
+                $this->query[self::QUERY_WHERE] .= $column . ' LIKE ?';
                 break;
 
-            case '!ew':
+            case self::OPERATOR_DOES_NOT_END_WITH:
 
                 if ($this->is_function($value)) {
-                    $this->query['where'] .= $column . ' NOT LIKE ' . $value;
+                    $this->query[self::QUERY_WHERE] .= $column . ' NOT LIKE ' . $value;
                     break;
                 }
 
                 $this->placeholders[] = '%' . $value;
-                $this->query['where'] .= $column . ' NOT LIKE ?';
+                $this->query[self::QUERY_WHERE] .= $column . ' NOT LIKE ?';
                 break;
 
-            case 'has':
+            case self::OPERATOR_HAS:
 
                 if ($this->is_function($value)) {
-                    $this->query['where'] .= $column . ' LIKE ' . $value;
+                    $this->query[self::QUERY_WHERE] .= $column . ' LIKE ' . $value;
                     break;
                 }
 
                 $this->placeholders[] = '%' . $value . '%';
-                $this->query['where'] .= $column . ' LIKE ?';
+                $this->query[self::QUERY_WHERE] .= $column . ' LIKE ?';
                 break;
 
-            case '!has':
+            case self::OPERATOR_DOES_NOT_HAVE:
 
                 if ($this->is_function($value)) {
-                    $this->query['where'] .= $column . ' NOT LIKE ' . $value;
+                    $this->query[self::QUERY_WHERE] .= $column . ' NOT LIKE ' . $value;
                     break;
                 }
 
                 $this->placeholders[] = '%' . $value . '%';
-                $this->query['where'] .= $column . ' NOT LIKE ?';
+                $this->query[self::QUERY_WHERE] .= $column . ' NOT LIKE ?';
                 break;
 
-            case 'in':
+            case self::OPERATOR_IN:
 
                 if ($this->is_function($value)) {
-                    $this->query['where'] .= $column . ' IN (' . $value . ')';
+                    $this->query[self::QUERY_WHERE] .= $column . ' IN (' . $value . ')';
                     break;
                 }
 
@@ -411,14 +441,14 @@ class Query
 
                 }
 
-                $this->query['where'] .= $column . ' IN (' . $in . ')';
+                $this->query[self::QUERY_WHERE] .= $column . ' IN (' . $in . ')';
 
                 break;
 
-            case '!in':
+            case self::OPERATOR_NOT_IN:
 
                 if ($this->is_function($value)) {
-                    $this->query['where'] .= $column . ' NOT IN (' . $value . ')';
+                    $this->query[self::QUERY_WHERE] .= $column . ' NOT IN (' . $value . ')';
                     break;
                 }
 
@@ -432,23 +462,23 @@ class Query
 
                 }
 
-                $this->query['where'] .= $column . ' NOT IN (' . $in . ')';
+                $this->query[self::QUERY_WHERE] .= $column . ' NOT IN (' . $in . ')';
 
                 break;
 
-            case 'null':
+            case self::OPERATOR_NULL:
 
-                if ($value == 'true') {
+                if ($value == self::VALUE_TRUE) {
 
-                    $this->query['where'] .= $column . ' IS NULL';
+                    $this->query[self::QUERY_WHERE] .= $column . ' IS NULL';
 
-                } else if ($value == 'false') {
+                } else if ($value == self::VALUE_FALSE) {
 
-                    $this->query['where'] .= $column . ' IS NOT NULL';
+                    $this->query[self::QUERY_WHERE] .= $column . ' IS NOT NULL';
 
                 } else {
 
-                    throw new QueryException('Unable to build query: invalid value (' . $value . ') for operator (null)');
+                    throw new QueryException('Unable to build query: invalid value (' . $value . ') for operator (' . self::OPERATOR_NULL . ')');
 
                 }
 
@@ -458,16 +488,16 @@ class Query
 
                 if ($value == '') { // Empty string needs no placeholder
 
-                    $this->query['where'] .= $column . " " . $operator . " ''";
+                    $this->query[self::QUERY_WHERE] .= $column . " " . $operator . " ''";
 
                 } else if ($this->is_function($value)) {
 
-                    $this->query['where'] .= $column . ' ' . $operator . ' ' . $value;
+                    $this->query[self::QUERY_WHERE] .= $column . ' ' . $operator . ' ' . $value;
 
                 } else {
 
                     $this->placeholders[] = $value;
-                    $this->query['where'] .= $column . ' ' . $operator . ' ?';
+                    $this->query[self::QUERY_WHERE] .= $column . ' ' . $operator . ' ?';
 
                 }
 
@@ -483,7 +513,7 @@ class Query
      * Values in the $columns array without a prefix or prefixed with a "+" will be ordered ascending.
      * Values in the $columns array prefixed with a "-" will be ordered descending.
      *
-     * If the column type is json, keys from within the JSON string can be ordered with the format of COLUMN->KEY.
+     * If the column type is JSON, keys from within the JSON string can be ordered with the format of COLUMN->KEY.
      * JSON fields which do not exist are treated as null.
      *
      * @param array $columns
@@ -525,7 +555,7 @@ class Query
 
         }
 
-        $this->query['sort'] = rtrim($string, ', ');
+        $this->query[self::QUERY_SORT] = rtrim($string, ', ');
 
         return $this;
 
@@ -538,7 +568,7 @@ class Query
      */
     public function orderByRand(): self
     {
-        $this->query['sort'] = ' ORDER BY RAND()';
+        $this->query[self::QUERY_SORT] = ' ORDER BY RAND()';
         return $this;
     }
 
@@ -550,7 +580,7 @@ class Query
      */
     public function limit(int $limit): self
     {
-        $this->query['limit'] = ' LIMIT ' . $limit;
+        $this->query[self::QUERY_LIMIT] = ' LIMIT ' . $limit;
         return $this;
     }
 
@@ -562,7 +592,7 @@ class Query
      */
     public function offset(int $offset): self
     {
-        $this->query['offset'] = ' OFFSET ' . $offset;
+        $this->query[self::QUERY_OFFSET] = ' OFFSET ' . $offset;
         return $this;
     }
 
@@ -574,16 +604,16 @@ class Query
     protected function getQuery(): string
     {
 
-        return 'SELECT ' . Arr::get($this->query, 'distinct', '')
-            . implode(', ', Arr::get($this->query, 'columns', []))
-            . Arr::get($this->query, 'from', '')
-            . Arr::get($this->query, 'inner_join', '')
-            . Arr::get($this->query, 'left_join', '')
-            . Arr::get($this->query, 'right_join', '')
-            . Arr::get($this->query, 'where', '')
-            . Arr::get($this->query, 'sort', '')
-            . Arr::get($this->query, 'limit', '')
-            . Arr::get($this->query, 'offset', '');
+        return 'SELECT ' . Arr::get($this->query, self::QUERY_DISTINCT, '')
+            . implode(', ', Arr::get($this->query, self::QUERY_COLUMNS, []))
+            . Arr::get($this->query, self::QUERY_FROM, '')
+            . Arr::get($this->query, self::QUERY_INNER_JOIN, '')
+            . Arr::get($this->query, self::QUERY_LEFT_JOIN, '')
+            . Arr::get($this->query, self::QUERY_RIGHT_JOIN, '')
+            . Arr::get($this->query, self::QUERY_WHERE, '')
+            . Arr::get($this->query, self::QUERY_SORT, '')
+            . Arr::get($this->query, self::QUERY_LIMIT, '')
+            . Arr::get($this->query, self::QUERY_OFFSET, '');
 
     }
 
@@ -704,11 +734,11 @@ class Query
     public function getTotalRows(): int
     {
         $query = 'SELECT COUNT(*)'
-            . Arr::get($this->query, 'from', '')
-            . Arr::get($this->query, 'inner_join', '')
-            . Arr::get($this->query, 'left_join', '')
-            . Arr::get($this->query, 'right_join', '')
-            . Arr::get($this->query, 'where', '');
+            . Arr::get($this->query, self::QUERY_FROM, '')
+            . Arr::get($this->query, self::QUERY_INNER_JOIN, '')
+            . Arr::get($this->query, self::QUERY_LEFT_JOIN, '')
+            . Arr::get($this->query, self::QUERY_RIGHT_JOIN, '')
+            . Arr::get($this->query, self::QUERY_WHERE, '');
 
         $stmt = $this->pdo->prepare($query);